Migrate APIs into c_api; Set up unit test framework.
