+++ tiledb/sm/c_api/tiledb_experimental.h
@@ -46,6 +46,7 @@
 #include "tiledb/api/c_api/attribute/attribute_api_external_experimental.h"
 #include "tiledb/api/c_api/current_domain/current_domain_api_external_experimental.h"
 #include "tiledb/api/c_api/enumeration/enumeration_api_experimental.h"
+#include "tiledb/api/c_api/fragment_info/fragment_info_api_experimental.h"
 #include "tiledb/api/c_api/query_aggregate/query_aggregate_api_external_experimental.h"
 #include "tiledb/api/c_api/query_field/query_field_api_external_experimental.h"
 #include "tiledb/api/c_api/query_plan/query_plan_api_external_experimental.h"
@@ -742,42 +743,6 @@ TILEDB_EXPORT int32_t tiledb_mime_type_to_str(
  */
 TILEDB_EXPORT int32_t tiledb_mime_type_from_str(
     const char* str, tiledb_mime_type_t* mime_type) TILEDB_NOEXCEPT;
-
-/**
- * Retrieves the number of cells written to the fragments by the user.
- *
- * Contributions from each fragment to the total are as described in following.
- *
- * In the case of sparse fragments, this is the number of non-empty
- * cells in the fragment.
- *
- * In the case of dense fragments, TileDB may add fill
- * values to populate partially populated tiles. Those fill values
- * are counted in the returned number of cells. In other words,
- * the cell number is derived from the number of *integral* tiles
- * written in the file.
- *
- * note: The count returned is the cumulative total of cells
- * written to all fragments in the current fragment_info entity,
- * i.e. count may effectively include multiples for any cells that
- * may be overlapping across the various fragments.
- *
- * **Example:**
- *
- * @code{.c}
- * uint64_t cell_num;
- * tiledb_fragment_info_get_total_cell_num(ctx, fragment_info, &cell_num);
- * @endcode
- *
- * @param[in]  ctx The TileDB context
- * @param[in]  fragment_info The fragment info object.
- * @param[out] count The number of cells to be retrieved.
- * @return `TILEDB_OK` for success and `TILEDB_ERR` for error.
- */
-TILEDB_EXPORT int32_t tiledb_fragment_info_get_total_cell_num(
-    tiledb_ctx_t* ctx,
-    tiledb_fragment_info_t* fragment_info,
-    uint64_t* count) TILEDB_NOEXCEPT;
 
 /**
  * Creates a consolidation plan object.
